# FrontEnd/文本生成.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "文本生成",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/graph/{eventGraphId}/data": {
      "get": {
        "summary": "获取指定事件图的所有信息",
        "deprecated": false,
        "description": "获取指定事件图的所有节点和关系",
        "operationId": "getEventGraphData",
        "tags": [
          "事件图节点和关系"
        ],
        "parameters": [
          {
            "name": "eventGraphId",
            "in": "path",
            "description": "事件图ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEventGraphDataDTO"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/text_styles/3": {
      "get": {
        "summary": "获取单个风格",
        "deprecated": false,
        "description": "获取指定风格的详细信息",
        "operationId": "getStyleById",
        "tags": [
          "文本风格"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TextStyle"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/deepseek/generate": {
      "post": {
        "summary": "生成文本",
        "deprecated": false,
        "description": "调用 DeepSeek API 生成文本",
        "operationId": "generateText",
        "tags": [
          "DeepSeek API"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/generated_texts": {
      "post": {
        "summary": "根据事件图和风格生成文本",
        "deprecated": false,
        "description": "通过指定事件图和风格生成新的文本",
        "operationId": "generateText",
        "tags": [
          "生成文本"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedText"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "查询指定事件图的所有文本",
        "deprecated": false,
        "description": "返回按创建时间降序排序的全部文本",
        "operationId": "getAllTextsByEventGraphId",
        "tags": [
          "生成文本"
        ],
        "parameters": [
          {
            "name": "eventGraphId",
            "in": "query",
            "description": "事件图ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeneratedText"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/generated_texts/{id}": {
      "put": {
        "summary": "编辑文本内容",
        "deprecated": false,
        "description": "修改指定ID的文本内容",
        "operationId": "updateText",
        "tags": [
          "生成文本"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "文本ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除指定文本",
        "deprecated": false,
        "description": "删除某条生成文本（仅限当前用户拥有的）",
        "operationId": "deleteText",
        "tags": [
          "生成文本"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "文本ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/generated_texts/latest": {
      "get": {
        "summary": "查询指定事件图的最近一条生成文本",
        "deprecated": false,
        "description": "获取指定事件图的最新生成文本",
        "operationId": "getLatestText",
        "tags": [
          "生成文本"
        ],
        "parameters": [
          {
            "name": "eventGraphId",
            "in": "query",
            "description": "事件图ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/generated_texts/{id}/export": {
      "get": {
        "summary": "导出单个文本",
        "deprecated": false,
        "description": "以.txt文件形式导出该文本",
        "operationId": "exportText",
        "tags": [
          "生成文本"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "文本ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "byte"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/generated_texts/export": {
      "get": {
        "summary": "导出指定事件图的所有文本",
        "deprecated": false,
        "description": "将该图下所有文本打包导出为zip文件",
        "operationId": "exportTextsByEventGraphId",
        "tags": [
          "生成文本"
        ],
        "parameters": [
          {
            "name": "eventGraphId",
            "in": "query",
            "description": "事件图ID",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "byte"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/generated_texts/export/all": {
      "get": {
        "summary": "导出用户的所有文本",
        "deprecated": false,
        "description": "将当前用户所有文本统一打包导出为zip文件",
        "operationId": "exportAllTexts",
        "tags": [
          "生成文本"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token2}}",
            "schema": {
              "type": "string",
              "default": "{{token2}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "byte"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "EventNodeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "eventGraphId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "labels": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        }
      },
      "TextStyle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "风格ID",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "风格名称",
            "examples": [
              "学术"
            ]
          },
          "description": {
            "type": "string",
            "description": "风格描述",
            "examples": [
              "适用于学术论文的写作风格"
            ]
          }
        },
        "description": "文本风格"
      },
...
      "EventRelationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "eventGraphId": {
            "type": "integer",
            "format": "int64"
          },
          "sourceId": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRECEDES",
              "FOLLOWS",
              "CONCURRENT_WITH",
              "CAUSES",
              "RESULTS_IN",
              "COREFERS_TO",
              "HAS_SUBEVENT",
              "SUBEVENT_OF",
              "SIMILAR_TO"
            ]
          },
          "strength": {
            "type": "string",
            "enum": [
              "低",
              "中",
              "高"
            ]
          }
        }
      },
...
      "UpdateTextRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
      "GenerateRequest": {
        "type": "object",
        "properties": {
          "systemPrompt": {
            "type": "string"
          },
          "userMessage": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseEventGraphDataDTO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/EventGraphDataDTO"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "GenerateResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "GenerateTextRequest": {
        "type": "object",
        "properties": {
          "eventGraphId": {
            "type": "integer",
            "format": "int64"
          },
          "styleId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "EventGraphDataDTO": {
        "type": "object",
        "properties": {
          "eventGraphId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventNodeDTO"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventRelationDTO"
            }
          }
        }
      },
      "GeneratedText": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "文本ID",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "eventGraphId": {
            "type": "integer",
            "description": "事件图ID",
            "format": "int64"
          },
          "styleId": {
            "type": "integer",
            "description": "风格ID",
            "format": "int64"
          },
          "content": {
            "type": "string",
            "description": "文本内容"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "生成的文本"
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}